+++ deploy/common/settings.py.j2
@@ -180,7 +180,7 @@ GW_BBH_MAX_AREA        = 300
 proxies = {}
 #-----------
 proxies['BBB_fast_SN'] = {
-    'CODE': 'filter.BBB_fast_SN',
+    'CODE': 'lasair_filter.BBB_fast_SN',
     'TOPIC': 'lasair_82BBBfastriserSN',
 }
 #-----------
